+++ TMF-source/slimevoid/tmf/items/ItemMineralMixedDust.java
@@ -39,7 +39,7 @@ public class ItemMineralMixedDust extends ItemMineralDust {
 		if ( speedLevel == 0 )
 			return 200;
 
-		double out = ( (Math.log((double)(speedLevel)+1d)-133d) * -(1d / ((double)(speedLevel)+1d) ) + 32d );
+		double out = ( (Math.log((speedLevel)+1d)-133d) * -(1d / ((speedLevel)+1d) ) + 32d );
 		//double out = 218.75d + ((double)(speedLevel+1)*18.75d);
 		if ( out < 50)
 			out = 50;
